apps/python-backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database - Tortoise ORM
tortoise-orm==0.21.7
aerich==0.7.2
aiosqlite==0.20.0

# Validation
pydantic==2.10.5
pydantic-settings==2.7.1

# CORS
python-multipart==0.0.20

# Development
pylint==3.3.3
mypy==1.14.1
black==24.10.0
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1

